+++ Assets/Script/Database/Table/CoordinateTable.cs
@@ -2,8 +2,8 @@
 {
     class CoordinateTable
     {
-        public float CoordinateY { get; set; }
-        public float CoordinateX { get; set; }
+        public double CoordinateY { get; set; }
+        public double CoordinateX { get; set; }
         public string Name { get; set; }
         public long Id { get; set; }
     }
